Add a public API
No breaking changes. An additional function exposed for package dependencies installation.

close #125
