+++ lib/api/init_cmd.js
@@ -14,11 +14,13 @@ var storeJsonController = require('../fs/store_json_controller')
 var mkdirp = require('../fs/mkdirp')
 
 module.exports = function (opts) {
+  opts = opts || {}
+  const cwd = opts.cwd || process.cwd()
   var cmd = {
     ctx: {}
   }
   var lockfile
-  return (opts.global ? readGlobalPkg(opts.globalPath) : readPkgUp())
+  return (opts.global ? readGlobalPkg(opts.globalPath) : readPkgUp({ cwd }))
     .then(_ => { cmd.pkg = _ })
     .then(_ => updateContext())
     .then(_ => mkdirp(cmd.ctx.store))
@@ -26,7 +28,7 @@ module.exports = function (opts) {
     .then(_ => cmd)
 
   function updateContext () {
-    var root = cmd.pkg.path ? dirname(cmd.pkg.path) : process.cwd()
+    var root = cmd.pkg.path ? dirname(cmd.pkg.path) : cwd
     cmd.ctx.root = root
     cmd.ctx.store = resolveStorePath(opts.storePath)
     lockfile = resolve(cmd.ctx.store, 'lock')
