+++ lib/install_multiple.js
@@ -5,13 +5,12 @@ var pkgFullName = require('./pkg_full_name')
  * Install multiple modules into `modules`.
  *
  *     ctx = { }
- *     installMultiple(ctx, ['minimatch@2'], ['chokidar@1.6.0'], './node_modules')
  *     installMultiple(ctx, { minimatch: '^2.0.0' }, {chokidar: '^1.6.0'}, './node_modules')
  */
 
 module.exports = function installMultiple (ctx, requiredPkgsMap, optionalPkgsMap, modules, options) {
-  requiredPkgsMap = mapify(requiredPkgsMap)
-  optionalPkgsMap = mapify(optionalPkgsMap)
+  requiredPkgsMap = requiredPkgsMap || {}
+  optionalPkgsMap = optionalPkgsMap || {}
 
   var optionalPkgs = Object.keys(optionalPkgsMap)
     .map(pkgName => pkgMeta(pkgName, optionalPkgsMap[pkgName], true))
@@ -46,22 +45,6 @@ module.exports = function installMultiple (ctx, requiredPkgsMap, optionalPkgsMap
         throw err
       })
   }))
-}
-
-function mapify (pkgs) {
-  if (!pkgs) return {}
-  if (Array.isArray(pkgs)) {
-    return pkgs.reduce((pkgsMap, pkgFullName) => {
-      var matches = /(@?[^@]+)@(.*)/.exec(pkgFullName)
-      if (!matches) {
-        pkgsMap[pkgFullName] = null
-      } else {
-        pkgsMap[matches[1]] = matches[2]
-      }
-      return pkgsMap
-    }, {})
-  }
-  return pkgs
 }
 
 function pkgMeta (name, version, optional) {
